voice-agent/requirements.txt
# ─── LiveKit Agent SDK ────────────────────────────────
livekit-agents>=1.5.0
livekit-plugins-openai>=0.10.0
livekit-api>=0.7.0

# ─── Sarvam AI (official LiveKit plugin) ──────────────
livekit-plugins-sarvam>=1.4.0

# ─── Silero VAD (Voice Activity Detection) ────────────
livekit-plugins-silero>=0.10.0

# ─── Web server (outbound call trigger) ──────────────
fastapi>=0.115.0
uvicorn>=0.32.0

# ─── Utilities ────────────────────────────────────────
python-dotenv>=1.0.0
aiohttp>=3.10.0
